requirements.txt
pandas
numpy
python-binance
pandas-ta
backtrader
tqdm
requests
python-dotenv
